+++ src/app/layout/AppLayout.tsx
@@ -3,7 +3,7 @@ import { useEffect, useState } from 'react';
 import { Footer } from '@/widgets/footer';
 import { Separator } from '@/shared/ui/shadcn/separator';
 import { SidebarProvider } from '@/shared/ui/shadcn/sidebar';
-import { SideBar } from '@/widgets/sidebar';
+import { SideBar } from '@/widgets/Sidebar';
 import { HeaderGuest, HeaderLogin } from '@/widgets/header';
 
 export default function AppLayout() {
