+++ src/entities/post/index.ts
@@ -2,3 +2,5 @@ export * from './model/queries';
 export * from './model/types';
 export * from './api/getPostsByCategory';
 export * from './ui/PostMeta';
+export * from './model/usePostsLandingPageCategoryQuery';
+export * from './model/usePostsHeartCategoryQuery';
