--- src/features/filter-posts/index.ts
@@ -1,1 +1,0 @@
-export * from './model/useCategoryFilter';
