+++ src/features/landing-page-filter-post/model/useFilter.ts
@@ -1,5 +1,5 @@
 import { useState } from 'react';
-import type { Category } from '@/features/filter-posts/api/getCategories';
+import type { Category } from '@/features/search-posts/api/getCategories';
 
 export function useFilter(initial: Category['code'] = 'ALL') {
   const [category, setCategory] = useState<Category['code']>(initial);
