+++ src/features/search-posts/api/getCategories.ts
