+++ src/features/search-posts/lib/useCategoriesQuery.ts
