+++ src/features/search-posts/model/queryKeys.ts
@@ -1,4 +1,4 @@
-import type { Params } from '@/widgets/PostInfo/model/type';
+import type { Params } from '@/widgets/PostList/model/type';
 
 export const postListKeys = {
   all: ['posts'] as const,
