--- src/widgets/PostInfo/index.ts
@@ -1,1 +1,0 @@
-export * from './ui/PostInfo';
