--- src/widgets/PostList/model/interface.ts
@@ -1,10 +1,0 @@
-import type { CategoryCode } from '@/entities/post/model/types';
-
-export interface PostListProps {
-  className?: string;
-  code?: CategoryCode;
-  page?: number;
-  size?: number;
-  limit?: number;
-  showPagination?: boolean;
-}
