+++ src/widgets/PostList/ui/PostCard.tsx
